working mvp

+++ DateTimeService/ResponseElement.cs
@@ -9,7 +9,7 @@ namespace DateTimeService
     {
         public string Nomenclature_id { get; set; }
 
-        public int Max_available_count { get; set; }
+        public decimal Max_available_count { get; set; }
 
     }
 }
